refactor(android): improve code structure
