+++ android/src/main/kotlin/ch/waio/pro_video_editor/src/core/constants/LoggingConstants.kt
@@ -1,1 +1,2 @@
 const val PACKAGE_TAG = "ProVideoEditor"
+const val RENDER_TAG = "$PACKAGE_TAG-Renderer"
